cloud_cost_optimizer_agent/agent.py
from google.adk.agents import Agent

root_agent = Agent(
    name="cloud_cost_optimizer_agent",
    model="gemini-2.5-flash",
    description="Helps optimize Google Cloud costs",
    instruction="""
You are a Google Cloud Cost Optimization expert.

Help users reduce cloud costs by suggesting:
- Rightsizing compute instances
- Using committed use discounts
- Using spot/preemptible VMs
- Storage lifecycle policies
- Autoscaling
- Monitoring idle resources
"""
)