cloud_cost_optimizer_agent/requirements.txt
google-adk==1.14.0
langchain-community==0.3.27
python-dotenv
google-cloud-logging
wikipedia